解决订单提交失败问题

+++ app/src/main/java/pro/haichuang/tzs144/model/SaleDataModel.java
@@ -118,7 +118,7 @@ public class SaleDataModel {
             this.type = type;
         }
 
-        public double getAddressId() {
+        public int getAddressId() {
             return addressId;
         }
